admin_page/requirements.txt
streamlit==1.39.0
streamlit-server-state==0.17.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit-webrtc==0.47.9
streamlit-modal==0.1.2
firebase-admin==6.6.0
Pyrebase4==4.8.0